Add: parseXML option in webdav
