+++ lib/drive/webdav.js
@@ -1,33 +1,11 @@
 /* global self */
-const {dirname} = require("path");
+const {dirname, basename} = require("path");
 
 const {createRequest} = require("../request");
+const {createXMLParser} = require("../util");
 
 function arrayify(o) {
   return Array.isArray(o) ? o : [o];
-}
-
-function xmlToJSON(node) {
-  // FIXME: xmldom doesn't support children
-  const children = Array.prototype.filter.call(node.childNodes, i => i.nodeType === 1);
-  if (!children.length) {
-    return node.textContent;
-  }
-  
-  const o = {};
-  for (const c of children) {
-    const cResult = xmlToJSON(c);
-    if (!o[c.localName]) {
-      o[c.localName] = cResult;
-    } else if (!Array.isArray(o[c.localName])) {
-      const list = [o[c.localName]];
-      list.push(cResult);
-      o[c.localName] = list;
-    } else {
-      o[c.localName].push(cResult);
-    }
-  }
-  return o;
 }
 
 function createDrive({
@@ -36,6 +14,7 @@ function createDrive({
   url,
   fetch = (typeof self !== "undefined" ? self : global).fetch,
   DOMParser = (typeof self !== "undefined" ? self : global).DOMParser,
+  parseXML = createXMLParser(DOMParser)
 }) {
   if (!url.endsWith("/")) {
     url += "/";
@@ -63,10 +42,7 @@ function createDrive({
       ...args
     });
     if (args.format || typeof text !== "string" || !text) return text;
-    
-    const parser = new DOMParser;
-    const xml = parser.parseFromString(text, "application/xml");
-    const result = xmlToJSON(xml);
+    const result = await parseXML(text);
     if (result.error) {
       throw new Error(`Failed requesting DAV at ${url}${path}: ${JSON.stringify(result.error)}`);
     }
@@ -104,9 +80,11 @@ function createDrive({
       if (arrayify(entry.propstat).some(s => s.prop.resourcetype && s.prop.resourcetype.collection !== undefined)) {
         continue;
       }
+      // FIXME: what is the proper way to get the filename from entry?
+      // NOTE: some servers may treat `@` and `%40` as the same path, we may have to normalize them here and there.
       const base = `${url}${file}`;
       const absUrl = new URL(entry.href, base).href;
-      const name = absUrl.slice(base.length);
+      const name = basename(absUrl);
       files.push(name);
     }
     return files;
